Fixed a bug that caused fread() to sometimes read too much data from a buffered stream

+++ Library/libc/Sources/stdio/__stdio.h
@@ -201,7 +201,7 @@ extern off_t __fgetlogicalpos(FILE * _Nonnull s);
 extern int __fseeko(FILE * _Nonnull s, off_t offset, int whence);
 extern int __fsetdir(FILE * _Nonnull s, int dir);
 extern void __fdiscard(FILE * _Nonnull s);
-extern int __ffill(FILE * _Nonnull s);
+extern ssize_t __ffill(FILE * _Nonnull s);
 extern int __fflush(FILE* _Nonnull s);
 extern int __setvbuf(FILE * _Restrict s, char * _Restrict buffer, int mode, size_t size);
 extern int __fclose(FILE* _Nonnull s);
